Completed step 2

+++ blockhouse/app/models.py
@@ -5,10 +5,10 @@ from django.db import models
 class StockData(models.Model):
     symbol = models.CharField(max_length=10)
     timestamp = models.DateTimeField()
-    open_price = models.DecimalField(max_digits=10, decimal_places=2)
-    close_price = models.DecimalField(max_digits=10, decimal_places=2)
-    high_price = models.DecimalField(max_digits=10, decimal_places=2)
-    low_price = models.DecimalField(max_digits=10, decimal_places=2)
+    open_price = models.FloatField()
+    close_price = models.FloatField()
+    high_price = models.FloatField()
+    low_price = models.FloatField()
     volume = models.BigIntegerField()
 
     class Meta:
